adds job tasks and tests
