+++ docs/jobs.md
@@ -1,0 +1,1 @@
+:::prefect_kubernetes.jobs
