--- docs/pod.md
@@ -1,1 +1,0 @@
-:::prefect_kubernetes.pod
