+++ prefect_kubernetes/exceptions.py
@@ -1,13 +1,15 @@
+"""Module to define common exceptions within `prefect_kubernetes`."""
+
 from kubernetes.client.exceptions import ApiException
 
 
 class KubernetesJobDefinitionError(Exception):
-    """An exception to raise when a Kubernetes job definition is invalid"""
+    """An exception for when a Kubernetes job definition is invalid"""
 
 
 class KubernetesJobFailedError(Exception):
-    """An exception to raise when a job orchestrated by a prefect_kubernetes task fails"""
+    """An exception for when a Kubernetes job fails"""
 
 
 class KubernetesResourceNotFoundError(ApiException):
-    """An exception to raise when a Kubernetes resource cannot be found by an api client"""
+    """An exception for when a Kubernetes resource cannot be found by a client"""
