+++ tests/test_jobs.py
@@ -1,14 +1,13 @@
 import pytest
 from kubernetes.client.exceptions import ApiValueError
 
-from prefect_kubernetes.job import (
+from prefect_kubernetes.jobs import (
     create_namespaced_job,
     delete_namespaced_job,
     list_namespaced_job,
     patch_namespaced_job,
     read_namespaced_job,
     replace_namespaced_job,
-    run_namespaced_job,
 )
 
 
@@ -90,30 +89,3 @@ async def test_replace_namespaced_job(kubernetes_credentials, api_batch_client):
     assert api_batch_client.replace_namespaced_job.call_args[1]["namespace"] == "ns"
     assert api_batch_client.replace_namespaced_job.call_args[1]["body"] == {"test": "a"}
     assert api_batch_client.replace_namespaced_job.call_args[1]["a"] == "test"
-
-
-# async def test_run_namespaced_job(
-#     kubernetes_credentials, api_batch_client, api_core_client
-# ):
-#     @flow
-#     def test_flow():
-#         run_namespaced_job(
-#             body={"metadata": {"name": "test"}},
-#             namespace="ns",
-#             kube_kwargs={"a": "test"},
-#             kubernetes_credentials=kubernetes_credentials,
-#             log_level="DEBUG",
-#         )
-
-#     test_flow()
-
-#     assert api_batch_client.create_namespaced_job.call_args[1]["body"] == {"test": "a"}
-#     assert api_batch_client.create_namespaced_job.call_args[1]["namespace"] == "ns"
-#     assert api_batch_client.create_namespaced_job.call_args[1]["a"] == "test"
-
-#     assert api_batch_client.read_namespaced_job_status.call_args[1]["body"] == {
-#         "test": "a"
-#     }
-#     assert api_batch_client.read_namespaced_job_status.call_args[1]["namespace"] == "ns"
-
-#     assert api_core_client.list_namespaced_pod.call_args[1]["namespace"] == "ns"
